+++ packages/@agla-utils/ag-logger/src/__tests__/units/utils/AgLogValidators.spec.ts
@@ -10,7 +10,7 @@ import { describe, expect, it } from 'vitest';
 import { AG_LOGLEVEL } from '../../../../shared/types';
 import type { AgLogMessage } from '../../../../shared/types';
 import { AgLoggerError } from '../../../../shared/types/AgLoggerError.types';
-import type { AgFormatRoutine } from '../../../internal/types/AgMockConstructor.class';
+import type { AgFormatRoutine } from '../../../../shared/types/AgMockConstructor.class';
 import { AgMockFormatter } from '../../../plugins/formatter/AgMockFormatter';
 import { isAgMockConstructor, isValidLogLevel, validateLogLevel } from '../../../utils/AgLogValidators';
